style login component with css
