+++ web-app/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { CryptoDetailsComponent } from './pages/cryptos/containers/crypto-details.component';
 import { authGuard } from './core/guards/auth.guard';
-import { LoginComponent } from './pages/auth/login.component';
+import { LoginComponent } from './pages/auth/login/login.component';
 import { ViewCryptosComponent } from './pages/cryptos/containers/view-cryptos/view-cryptos.component';
 import { roleGuard } from './core/guards/role.guard';
 import { AccountComponent } from './pages/cryptos/containers/account/account.component';
