+++ web-app/src/app/pages/auth/login/login.component.html
@@ -1,0 +1,107 @@
+<div class="login-container">
+  <div class="login-row">
+    <div class="login-col">
+      <form (ngSubmit)="login()" class="login-form" [formGroup]="loginForm">
+        <h2 class="login-title">DG Invest</h2>
+        <div class="form-group">
+          <label for="email">Email</label>
+          <input required formControlName="email" type="email" id="email" placeholder="Enter your email">
+          <div class="error-message"
+            *ngIf="emailFormControl?.hasError('email') && !emailFormControl?.hasError('required')">
+            Please enter a valid email address.
+          </div>
+        </div>
+        <div class="form-group">
+          <label for="password">Password</label>
+          <input required formControlName="password" type="password" id="password" placeholder="Enter your password">
+          <div class="error-message" *ngIf="passwordFormControl?.hasError('minlength')">
+            Minimum password length is 4 characters.
+          </div>
+        </div>
+        <div class="form-group text-center">
+          <button type="submit" class="login-btn" [disabled]="loading || loginForm.invalid"
+            [ngStyle]="{'cursor': loading ? 'not-allowed' : 'pointer'}">{{loading ? 'Loading...' : 'Login'}}</button>
+        </div>
+      </form>
+    </div>
+  </div>
+</div>
+
+<style>
+  .login-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    background-color: #f9f9f9;
+  }
+
+  .login-row {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+  }
+
+  .login-col {
+    max-width: 600px;
+    width: 100%;
+  }
+
+  .login-form {
+    padding: 32px;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    background-color: white;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+
+  .login-title {
+    margin-bottom: 24px;
+    text-align: center;
+    font-size: 24px;
+  }
+
+  .form-group {
+    margin-bottom: 16px;
+    display: flex;
+    flex-direction: column;
+  }
+
+  .form-group label {
+    margin-bottom: 8px;
+    font-weight: bold;
+  }
+
+  .form-group input {
+    padding: 8px;
+    font-size: 16px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+  }
+
+  .error-message {
+    color: red;
+    font-size: 14px;
+    margin-top: 8px;
+  }
+
+  .login-btn {
+    padding: 12px 16px;
+    font-size: 16px;
+    color: white;
+    background-color: #007bff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+
+  .login-btn:disabled {
+    background-color: #ccc;
+    cursor: not-allowed;
+  }
+
+  .text-center {
+    text-align: center;
+  }
+</style>
